SpinShuttling.jl

Simulate the multiple-spin shuttling problem under correlated stochastic noise.

Installation

SpinShuttling can be installed using the Julia package manager. From the Julia REPL, type ] to enter the Pkg REPL mode and run

pkg> add SpinShuttling

APIs

SpinShuttling.OneSpinModel — Function

General one spin shuttling model initialized at initial state |Ψ₀⟩, with arbitrary shuttling path x(t).

Arguments

Ψ::Vector{<:Number}: Initial state of the spin system, the length of the vector must be `2^n

T::Real: Maximum time

M::Int: Monte-Carlo sampling size

N::Int: Time discretization

B::GaussianRandomField: Noise field

x::Function: Shuttling path

One spin shuttling model initialzied at |Ψ₀⟩=|+⟩. The qubit is shuttled at constant velocity along the path x(t)=L/T*t, with total time T in μs and length L in μm.

SpinShuttling.TwoSpinModel — Function

General two spin shuttling model initialized at initial state |Ψ₀⟩, with arbitrary shuttling paths x₁(t), x₂(t).

Arguments

Ψ::Vector{<:Number}: Initial state of the spin system, the length of the vector must be `2^n

T::Real: Maximum time

M::Int: Monte-Carlo sampling size

N::Int: Time discretization

B::GaussianRandomField: Noise field

x₁::Function: Shuttling path for the first spin

x₂::Function: Shuttling path for the second spin

Two spin shuttling model initialized at the singlet state |Ψ₀⟩=1/√2(|↑↓⟩-|↓↑⟩). The qubits are shuttled at constant velocity along the path x₁(t)=L/T₁*t and x₂(t)=L/T₁*(t-T₀). The delay between the them is T₀ and the total shuttling time is T₁+T₀. It should be noticed that due to the exclusion of fermions, x₁(t) and x₂(t) cannot overlap.

SpinShuttling.fidelity — Function

Sample a phase integral of the process. The integrate of a random function should be obtained from directly summation without using high-order interpolation (Simpson or trapezoid).

SpinShuttling.sampling — Function

Monte-Carlo sampling of any objective function. The function must return Tuple{Real,Real} or Tuple{Vector{<:Real},Vector{<:Real}}

Sampling an observable that defines on a specific spin shuttling model objective(mode, randseq, vector)

SpinShuttling.averagefidelity — Function

Calculate the average fidelity of a spin shuttling model using numerical integration of the covariance matrix.

SpinShuttling.OrnsteinUhlenbeckField — Type

Ornstein-Uhlenbeck field, the correlation function of which is σ^2 * exp(-|t₁ - t₂|/θ_t) * exp(-|x₁-x₂|/θ_x) where t is time and x is position.

SpinShuttling.PinkBrownianField — Type

Pink-Brownian Field, the correlation function of which is σ^2 * (expinti(-γ[2]abs(t₁ - t₂)) - expinti(-γ[1]abs(t₁ - t₂)))/log(γ[2]/γ[1]) * exp(-|x₁-x₂|/θ) where expinti is the exponential integral function.

SpinShuttling.RandomFunction — Type

Similar type of RandomFunction in Mathematica. Generate a time series on a given time array subject to a Gaussian random process traced from a Gaussian random field.

SpinShuttling.Χ — Function

Theoretical fidelity of a sequenced two-spin EPR pair shuttling model.

Theoretical fidelity of a one-spin shuttling model.

Theoretical fidelity of a one-spin shuttling model for a pink noise.